requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
pydantic==2.9.2
python-dotenv==1.0.1
python-telegram-bot==21.6
starlette==0.38.0
aiohttp==3.9.1
web3==6.11.0
python-multipart==0.0.9
pydantic-settings==2.6.1